Rewrite tests as integration tests
