+++ Cargo.toml
@@ -27,7 +27,6 @@ openssl-sys = { version = "0.9.0", optional = true }
 openssl-probe = { version = "0.1", optional = true }
 
 [dev-dependencies]
-serial_test = "0.5.1"
 structopt = "0.3"
 time = "0.1.39"
 tempfile = "3.1.0"
